+++ kiexport.py
@@ -5,7 +5,7 @@
 # Tool to export manufacturing files from KiCad PCB projects.
 # Author: Vishnu Mohanan (@vishnumaiea, @vizmohanan)
 # Version: 0.0.2
-# Last Modified: +05:30 23:19:21 PM 29-08-2024, Thursday
+# Last Modified: +05:30 23:33:42 PM 29-08-2024, Thursday
 # GitHub: https://github.com/vishnumaiea/KiExport
 # License: MIT
 
@@ -13,6 +13,7 @@
 
 import subprocess
 import argparse
+import os
 
 #=============================================================================================#
 
@@ -24,6 +25,14 @@ def generateGerbers (output_dir, pcb_filename):
   # output_dir = "Mitayi-Pico-D1/Gerber"
   layer_list = "F.Cu,B.Cu,F.Paste,B.Paste,F.Silkscreen,B.Silkscreen,F.Mask,B.Mask,User.Drawings,User.Comments,Edge.Cuts,F.Courtyard,B.Courtyard,F.Fab,B.Fab"
   # pcb_filename = "Mitayi-Pico-D1/Mitayi-Pico-RP2040.kicad_pcb"
+
+  if not check_file_exists (pcb_filename):
+    print (f"generateGerbers [ERROR]: {pcb_filename} does not exist.")
+    return
+
+  file_name = extract_pcb_file_name (pcb_filename)
+  project_name = extract_project_name (file_name)
+  print (f"generateGerbers [INFO]: Project name is {project_name}.")
 
   full_command = gerber_export_command + \
                 ["--output", output_dir] + \
@@ -39,6 +48,46 @@ def generateGerbers (output_dir, pcb_filename):
     print ("generateGerbers [OK]: Gerber files exported successfully.")
   except subprocess.CalledProcessError as e:
     print (f"generateGerbers [ERROR]: Error occurred: {e}")
+
+#=============================================================================================#
+
+def check_file_exists (file_name):
+  """
+  Checks if the input PCB file exists.
+  Args:
+    input_file_name (str): The path to the PCB file.
+  Returns:
+    bool: True if the file exists, False otherwise.
+  """
+  return os.path.exists (file_name)
+
+#=============================================================================================#
+
+def extract_project_name (file_name):
+  """
+  Extracts the project name from a given PCB file name by removing the extension.
+  Args:
+    input_file_name (str): The PCB file name.
+  Returns:
+    str: The project name without the file extension.
+  """
+  file_name = extract_pcb_file_name (file_name)
+  project_name = os.path.splitext (file_name) [0]
+  # print ("Project name is ", project_name)
+
+  return project_name
+
+#=============================================================================================#
+
+def extract_pcb_file_name (file_name):
+  """
+  Extracts the PCB file name from a given path.
+  Args:
+    input_file_name (str): The path to the PCB file.
+  Returns:
+    str: The PCB file name without any directory path.
+  """
+  return os.path.basename (file_name)
 
 #=============================================================================================#