+++ Changes.md
@@ -1,6 +1,16 @@
 
 
 #
+### **+05:30 11:33:55 PM 29-08-2024, Thursday**
+
+  - Added `check_file_exists()` to check is a file exists.
+  - Added `extract_project_name()` to extract the project name from the file name. This simply removes the extension.
+  - Added `extract_pcb_file_name()` to extract the file name from the path.
+  - `generateGerbers()` now prints the KiCad PCB project name.
+  - Code running successfully.
+  - New Version `0.0.3`.
+
+#
 ### **+05:30 11:18:09 PM 29-08-2024, Thursday**
 
   - Added CLI argument parser.